+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG (v0.2.X)
 
-## 0.2.0 (:soon:)
+## 0.2.0 (2025-10-27)
 
 ### ⚠️ Backwards incompatible changes for 0.1.12